Create bulk export models
We would like to be able to export credentials, both for backup and
mass-editing purposes.  These bulk export models provide a mechanism to turn
a slice of `credhub.credentials.Credential` into a YAML representation that
can be serialised or reimported by the bulk import process.

A test ensures that whatever is created remains importable as the code evolves.