Add a command to export credentials
Exporting credentials is useful both for backup and management purposes
(e.g. mass-moving a number of credentials to a different path). This
command enables exporting by path matching, utilising the bulk export
models to create an output that is reimportable.

By default, the command exports all credentials to `stdout` in import
compatible YAML. The `-p` flag can be used to restrict the paths
exported; the `-f` flag to write the output to a file. Note that the
credentials are held unencrypted in-memory during this process; as this
seems true of all the code, this should not be seen as a problem; just a
caveat.