Don't print insecure HTTP error when CLI defaults to HTTPS
This commit fixes a bug reported to us earlier this week. When setting
the server domain to a scheme-less domain (i.e. `credhub api
myserver.org`), CredHub implicitly converts it to
`https://myserver.org`. However, logic within the CLI would still print
a warning mentioning the insecurity of an HTTP connection, despite the
fact that the connection was made over HTTPS. We've modified the warning
logic to ensure an implicit "cast" to HTTPS does not print HTTP-related
errors.

[#166967869]

Signed-off-by: Gemma Standley <gstandley@pivotal.io>