fix: export of child certificates
* When writing out child certificates on `export` the CLI would put the
name of the signer into the "ca" field for the credential. This caused
import to fail, since import expects that only the value of the CA will
be in that field.

```
Credential '/bruce/ca-leaf' at index n could not be set: The provided CA value is not a valid X509 certificate authority.
```

* We now put the name (path) of the signer in a "ca_name" field. This is
what the import logic expects and handles this case already.

[#174125828](https://www.pivotaltracker.com/story/show/174125828)

Co-authored-by: Bruce Ricard <bricard@vmware.com>