CLI support for client certificates (mutual TLS)
This change allows using the Credhub CLI in CF apps without having
to make separate roles. The CF Instance Identity certificates [1] in
each CF app instance can be used to obtain a `mtls-app:APP_GUID` role
in CredHub. This can be useful in things such as buildpacks.

The `credhub` library in the codebase already supported mutual TLS, so
this change has merely been a matter of extending support to the CLI.