Fix a mistaken test
This commit fixes a test that didn't make sense. The test said that
the configured HTTP timeout was written to disk, but then the test
code checked that the HTTP timeout wasn't being written to disk.

The only reason the test wasn't failing is that it was wrongly
converting a number to a string.