Output the duration used when it is overridden
Newer versions of the CredHub server will include two new fields -
duration_overridden, which indicates the server-configured minimum was
used instead of the requested duration, and duration_used, which
indicates the final duration used to generate / regenerate the
certificate.

When these fields are present, and the duration_overridden flag is true,
the CLI will print out:
duration_overridden_to: X

with the value of the actual duration used.

Co-authored-by: Preethi Varambally <pvarambally@pivotal.io>
Co-authored-by: Brian Upton <bupton@vmware.com>
Co-authored-by: Pablo Rodas <prodas@vmware.com>