refactor test: quote yaml timestamp string correctly
- In anticipation of CI's `update-go-modules` job,
which will upgrade go yaml lib which will result in some
test failures, because the new yaml lib would
add double quotes to our `version_created_at` value in our yaml output,
which is in timestamp format but we have been treated it as a string.
- Also, the other test expectations in the same test class
are already quoting `version_created_at` value in the test
expectations, so this commit makes this consistent in all
tests in the same test file.

[#182918401]

Co-authored-by: Peter Chen <peterch@vmware.com>