Remove redundant type assertions.
c.Value's type is already `interface{}`, so
explicitly checking to see if it can be type-asserted to the
type it has is redundant. Additionally,this results in c.Value
being stored to result.Value both when it is really a string, or
otherwise, so just set it and move on.